refactor: move updateBounceDoc into schema methods
